Issue #6: Tested both applications using container with all dependencies.
We created a single dependency file with all the required packages for WineGuard and DairyManager.

Note that, for DairyManager, the test used datasets that will be packaged with the self-contained container to be loaded by the evaluator(s).

This ensures that the datasets, which may potentially contain vendor names, are not uploaded as part of the branchwhich will be released.